Android_UI_Appium/requirements.txt
Appium-Python-Client==1.0.2
concurrent-log-handler
python-dateutil

